Considering one-per-household setting when calculating weights

+++ lib/local/sample/gps.cxx
@@ -15,7 +15,7 @@ namespace sample
   //-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-+#+-
   double gps::get_sample_weight( const sampsim::individual* individual ) const
   {
-    return sized_sample::get_sample_weight( individual ) * this->number_of_buildings;
+    return sized_sample::get_sample_weight( individual ) * static_cast< double >( this->number_of_buildings );
   }
 }
 }
